Finished Project 3 (Ranking Hospitals)
This project consisted of creating three function scripts to analyze the hospital data contained in the zip file. The best.R script returned a character vector stating the name of the hospital with the best outcome result given the two inputs of state name and health condition (heart attack, heart failure, or pneumonia). The rankhospital.R script returned a character vector of the name of the hospital meeting the given parameters of state, health condition, and a ranking of either best, worst, or a given numerical index. The rankall.R script performed nearly the same function as the rankhospital.R script but with the output being a dataframe returning a hospital and state name for every state given the parameters of health condition and ranking.